git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@10801 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/ANALYSIS/OPENSWATH/MRMFeatureFinderScoring.C
@@ -139,7 +139,7 @@ namespace OpenMS
     Size nr_chromatograms = input->getNrChromatograms();
     for (Size i = 0; i < input->getNrChromatograms(); i++)
     {
-      chromatogram_map[input->getChromatogramNativeID(i)] = i;
+      chromatogram_map[input->getChromatogramNativeID(i)] = boost::numeric_cast<int>(i);
     }
 
     // Iterate thorugh all transitions and store the transition with the
